RAR5 reader: fix an uninitialized variable
The `rar5_read_data` callback is being called by libarchive when it
needs the RAR5 reader to provide a new chunk of decompressed data. This
callback returns the data size, data offset and the data itself.

Libarchive expects both the "data size" and "data offset" to be updated
by the decompressor. Not touching the "data offset" in some cases
resulted in a branch based on uninitialized value. One of the scenarios
where this bug can happen is described in GitHub issue #1738.

This commit always initializes the "data offset" to some value, which
removes the uninitializes value bug.

The commits also adds a test for this case to guard against future
regressions.

Fixes #1738.